+++ CHANGELOG.md
@@ -13,6 +13,13 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v4.0.0-beta.?? (2024-??-??)
 
+#### :boom: Breaking Change
+
+* The modules `core/init` and `core/component/state` have been completely redesigned for the new API
+* The fields `isOnline`, `lastOnlineDate`, and `locale` have been removed `iStaticPage`
+
+## v4.0.0-beta.?? (2024-??-??)
+
 #### :bug: Bug Fix
 
 * The `activeStore` system field is marked as unique in the `bList`, `bTree`, `bSelect`